requirements.txt
psycopg2-binary>=2.9.0
sqlalchemy>=1.4.0
python-dotenv>=1.0.0
pandas>=1.5.0
openpyxl>=3.0.0
pyyaml>=6.0